// dashboard/src/lib/api/InitialState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AuthStatus } from "./AuthStatus";
import type { CircuitBreakerStatus } from "./CircuitBreakerStatus";
import type { Inventory } from "./Inventory";
import type { PerformanceMetrics } from "./PerformanceMetrics";
import type { RebalanceOperation } from "./RebalanceOperation";
import type { SpreadSummary } from "./SpreadSummary";
import type { Trade } from "./Trade";

export type InitialState = { recentTrades: Array<Trade>, inventory: Inventory, metrics: PerformanceMetrics, spreads: Array<SpreadSummary>, activeRebalances: Array<RebalanceOperation>, recentRebalances: Array<RebalanceOperation>, authStatus: AuthStatus, circuitBreaker: CircuitBreakerStatus, };